% Generated by roxygen2 (4.0.2): do not edit by hand
\name{theme_gr}
\alias{theme_gr}
\title{ggplot2 grtheme}
\usage{
theme_gr(base_size = 12, base_family = "sans")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
Themes set the general aspect of the plot such as the colour of the
background, gridlines, the size and colour of fonts.
}
\author{
Giuseppe Ragusa
}

